Update patch set 4
Patch Set 4:

Er you're right, it was correct before Nathan's comment.

I keep on getting mixed signals about frame-parallel. BBB, the author of ffvp9, said that the performance improvement from it was minimal. Current testing forces frame-parallel off.

IMO if we leave frame-parallel on, it should be not possible to turn it off in the encoder so that decoders can depend on its existence. I don't really have a strong opinion either way.

I'll fix this patch, and then we should separately discuss the default value. (are you on IRC?)

Patch-set: 4
Label: Code-Review=0