Update patch set 7
Patch Set 7: Code-Review+2

(6 comments)

I like the intent of this and hope it passes through the process and gets into the codec! It should make things faster!

That said stylistically this patch is a lot larger than it has to be.   

Perhaps its a discussion for the group.  

I think this could have been done by making a non bitstream change part that split the loopfilter functions into vertical and horizontal filters and then adding calls to those new functions in av1_loop_filter_rows and thus localized all the non bitstream changes to that function

This would make the review a lot more onerous than it is now.

Patch-set: 7