Update patch set 7
Patch Set 7:

> Er you're right, it was correct before Nathan's comment.
 > 
 > I keep on getting mixed signals about frame-parallel. BBB, the
 > author of ffvp9, said that the performance improvement from it was
 > minimal. Current testing forces frame-parallel off.
 > 

It varies based on resolution. I measured ~2-3.5% drop in libvpx and ffvp9 on various intel architectures, this likely would increase with resolution and that kind of drop is significant on mobile as well as low-end laptops.
As for density improvements this trails off rather quickly above 480p (even that isn't very impressive at < 1% across a large corpus).

 > IMO if we leave frame-parallel on, it should be not possible to
 > turn it off in the encoder so that decoders can depend on its
 > existence. I don't really have a strong opinion either way.
 > 
 > I'll fix this patch, and then we should separately discuss the
 > default value. (are you on IRC?)

I've been trying to stay in the background in this project. It's fine to decide to go either way, I was only trying to make the point that it should be data-driven based on the target use cases.

Patch-set: 7