Add assembly implementation of weighted prediction for obmc
Port vpx_blend_mask6() from libvpx/nextgenv2 to AV1. It is originally
designed for wedge inter prediction, in which two prediction blocks
are combined according to a mask block.
To reuse it for obmc,
(1) vertical filtering: we replicate the 1d mask in every column to
generate the mask block
(2) horizontal filtering: set the pointer to the mask block = the pointer
to the 1d mask, and set mask_stride = 0

Trivial change in encoding speed
Decoding speed: +20.1% -> +8.4%

Change-Id: Ib6d793bd39605cb141329499a4031ff7fa337b34