Update patch set 12
Patch Set 12:

(2 comments)

I thought one of the points of this is that in truth each row of loop filter doing vertical could run completely independently?   Ie no synchronization is required in that direction and we could do each row completely independently.

Conversely reordering so that the columns were handled in the outer loop would remove the need for synchronization in the horizontal loop.   

I suggest we remove the synchronization in the one we can remove it and add some TODOs  as per TODO(wenhao.zhang@intel.com): better parallelization of loop filter by reordering the loop to columns

Patch-set: 12
Label: Code-Review=0