Disable the unit test of ArfFreqTest for ext-refs.
The test in arf_freq assumes any no-show frame as ALTREF_FRAME and
then calculate the minimum run between two consecutive ALTREF_FRAME's
based on this assumption. As BWDREF_FRAME is also a no-show frame and
the minimum run between two consecutive BWDREF_FRAME's may vary
between 1 and any arbitrary positive number as long as it does not
exceed the golden frame group interval, this test does not apply to
the experiment of ext-refs. Therefore, ArfFreqTest is disabled when
ext-refs is on.

Change-Id: I970f58fbfb8d2a03160f089748dccf1ad24207ee