Encode and decode multiple tile groups.
A tile group is a set of tiles in scan order.

Each tile group has a version of uncompressed and compressed headers.
Encoding probability updates takes account of the number of
headers to control overheads.

The decoder supports arbitrary numbers of tile groups with
arbitrary number of tiles. The number of tiles in a TG is
signalled in the uncompressed header for that TG.

The encoder currently only supports a fixed number
of TGs (3, when error resilient mode is on) of equal size
(except possibly for the last one).

Change-Id: I9b97c5ed733103b9160a3a5d4370de5322c00c0b