Encode and decode multiple tile groups.
A tile group is a set of tiles in scan order.

Each tile group has a copy of uncompressed and compressed headers.
Encoding probability updates takes account of the number of
headers to control overheads.

Only a fixed number of tile groups in currently supported (defined
by macro).

TODO

- Parallel decoding with (or by) tile groups
- Make TILE_GROUPS compatible with MISC_FIXES experiment
- Make TILE_GROUPS compatible with EXT_REF experiment
- Allow a variable number of tile groups
- Specify tile groups by data size
- Use smaller number of bits for tile sizes and code in tile group
header

Change-Id: I9b97c5ed733103b9160a3a5d4370de5322c00c0b