Refactor RDO State Save / Restore
Removes PICK_MODE_CONTEXT and the need to reencode blocks at every level of
the partition tree. The tree of PICK_MODE_CONTEXTs is replaced with one
RDContext for each level of the partition tree, which saves the necessary
information about the best mode / partition size so far.

Currently this refactor achieves about equal performance to the pctree
implementation, however it reencodes substantially less, which will improve
the speed of PVQ.

TODO
 - only reencode at the end of rd_pick_partition / rd_use_partition after
   the partition type is selected. This should provide a substantional
   performance boost, and can possibly be dependent on a speed feature.
 - Avoid redundant copies and calls to set_offsets

Change-Id: Icd7a8b1b3b02769da39bec4bb89a6aa93574654b