Update patch set 6
Patch Set 6:

Yaowu,

I tried to rearrange functions a little so more code would line up in the diff, it helped a little on reducing the insertion / deletion count, but not much. The majority of the changes in encodeframe.c are just code moves, but lines had to move around enough that it's a little beyond git's ability to match code chunks.

Patch-set: 6
Label: Code-Review=0