Update patch set 3
Patch Set 3:

> This is actually intended. Sub-8x8 blocks are handled a little differently from regular (8x8 and larger) blocks. Take 4x4 sub-blocks in a 8x8 partition block as an example: w=2(instead of 1), step=1. For block 0 and 2, x=0, x+step=1<w=2, so has_right=true; for blocks 1 and 3, x=1, x+step=2=w, so has_right=false.

This may be intentional, but it is broken. Consider the case of two 4x8 (w x h) partitions. The bottom transform block in the first partition will be coded before the top transform block in the second partition, but has_right will be true. That's the bug this is trying to solve.

Patch-set: 3
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=0