Update patch set 8
Patch Set 8:

We tested the time on objective-1-fast and it gave about a 1% overall speed improvement, which is in line with my assertion that overall it doesn't change speed much. Perhaps there is more overhead with smaller videos that gets offset by the larger videos in the set?

Patch-set: 8