Update patch set 3
Patch Set 3:

> (4 comments)
 > 
 > This looks like an interesting approach but do you have any
 > comparison between this and maybe some of the existing sad or
 > variance functions in libvpx?

No, but we've been using this approach for a long time (a decade) within Cisco, and in most cases this approach gives identical or nearly identical code as if native intrinsics were used.  There will always be some kernels in a codec that will benefit significantly from using native code, but those would not make up the majority.  Sad and variance functions will not run much faster if written using native code on x86 or arm.  Note, though, that accumulation for such functions work a bit differently on x86 and arm, but the common intrinsics have been designed to work around this.  For the most part, vector instructions are identical across platforms.

The idea of this library is not to give optimal code, but to get 90% of the way for a relatively small amount of work.  Cisco has many video products in the market optimised this way.

Patch-set: 3