Update patch set 3
Patch Set 3:

(4 comments)

I'm not opposed to adding another optimization tool, but this doesn't seem to fit well (as implemented) with the other tools like intrinsics and x86inc.asm. Would it be possible to restructure this so it fits with the existing RTCD system?

I'm not opposed to offering another style for quickly writing optimizations, but could you add in a quick example, and tie it into the tests? the _sad_functions are simple and well tested, would it be possible to offer up an implementation of say the 16x16 so we can compare it to the existing options?

Patch-set: 3