Update patch set 1
Patch Set 1:

Previously the decoder would write the filtered frame to a new buffer, then copy that buffer back to the original frame.  This patch enables the clpf_frame function to do in-place filtering, and unfiltered blocks are not copied at all.  Since the input is always unfiltered pixels, the function must have an internal buffer.  For simplicity, this a circular buffer that is frame_width*filter_block_size.

Patch-set: 1