Only reencode each block once
This patch adds a new buffer, recon_pix, to the PICK_MODE_CONTEXT,
which stores the reconstructed pixels created by reencoding the block.
After reencoding a block once, the reconstructed pixels are saved to
this buffer. The quantized coefficients are already saved in the
PICK_MODE_CONTEXTs buffers, so these reconstructed pixels simply need
to be restored rather than repeatedly reencoding the block.

The change is bit exact on high latency mode on objective-1-fast:
https://arewecompressedyet.com/?r%5B%5D=master2016-08-25T01-38-40.728Z&r%5B%5D=rrencode2016-08-25T01-38-00.353Z&s=objective-1-fast

The change is bit exact on low latency mode on objective-1-fast:
https://arewecompressedyet.com/?r%5B%5D=master-8-2016-08-25T01-42-10.505Z&r%5B%5D=rreencode8-2016-08-25T01-39-10.706Z&s=objective-1-fast

In my testing this patch gives little performance change in high
latency mode, and about 2% speed improvement in low latency mode,
depending on the clip.

AQ modes 2 and 3 don't make use of this change, this issue is
addressed in the next patch.

Change-Id: I241643712ccac943ec08e858be2862e0e0f22105