Only reencode once for AQ modes 2 and 3
These AQ modes change the segment id of a block based on the
information stored in the global segmentation_map. This is a problem,
because the global map can be changed after the first reencode, so
subsequent reencodes would give different results (and cause an encode
mismatch when only reencoding once).  This patch fixes this issue by
only changing the segment_id for a block on the first reencode.

This shows there is a slight rate reduction on AQ mode 2 with this
patch on objective-1-fast:
https://arewecompressedyet.com/?r%5B%5D=r-aq-2-2016-08-25T02-06-46.814Z&r%5B%5D=m-aq-2-2016-08-25T02-03-58.490Z&s=objective-1-fast

This shows there is no change on AQ mode 3 with this patch:
https://arewecompressedyet.com/?r%5B%5D=m-aq-3-2016-08-25T02-05-54.944Z&r%5B%5D=r-aq-3-2016-08-25T02-04-41.370Z&s=objective-1-fast

Change-Id: I69748e9799398ef2e09d913f8d3c5717506d8a99