Update patch set 11
Patch Set 11:

(8 comments)

I disagree that the managing the borders in the filter can be done efficiently. While it's certainly much faster than the C, removing the extra code handling that would be way better. Alternatively, simpler solutions to handling the border may be possible. The A and F terms (and B, C, D, and E, for the left- and right-most X values) go to 0. Finding a way to simply avoid calculating them would be much preferable.

Patch-set: 11
Label: Code-Review=0