Update patch set 11
Patch Set 11:

(4 comments)

There is a test outside the block loop checking for left and right clipping, so the overhead is only done when there's a need.  This test could of course be move outside the function which instead could be split for each case, but the top and bottom clipping is trivial, so making special functions for that seems unnecesary.

Usually it's not worthwhile to optimise vectorised operations though 1 or 2 out of 8 operations evaluate to 0.  And even if it were, a suboptimal solution would only concern a small portion of the frame anyway.  I tend to favour a pragmatic approach.  I prefer the simple 95% optimal solution over the complicated 100% optimal solution.

Patch-set: 11
Label: Code-Review=0