Update patch set 15
Patch Set 15:

Thanks for the extra information and tests. I don't have a problem with this particular implementation of this function, my concerns are with the design of the filter.

I've added aconverse@ because I'm not sure where to send this feedback.

I don't like this filter from an optimization standpoint. Having bounds checking at this level is very expensive. Yes, you're working around it with conditionals and shuffling, but it would be substantially faster without those.

Best would be to remove the edge behavior entirely and have it padded/adjusted at a higher level.

Another level would be removing the calculations at the edge, instead of adjusting them. For the top and bottom (A and F in the design document), both terms are 0 at the edge and can be simply discarded. For the left and right values, next to the edge both B and C (D and E on the right) are discarded. If this could be pulled in to the design of the filter and simplified, that would help too.

Patch-set: 15
Label: Code-Review=0