Update patch set 17
Patch Set 17:

(16 comments)

PS16-17 removes a couple of redundant instructions, changes temp variable names, argument order, adds consts to temp variables, etc as per Johann's suggestions.  Also fixes a typo that broke the test_libaom build for NEON.

NEON speed-up of clpf_block measured compared to C version: 12.79x.

I did not add a test for the (pathological) case when the frame width is less than 16 pixels.  I'm not sure if that's allowed and not expanded internally.  The test doesn't cost much if there is a need for it, but I'll rather await further comments on that.

Patch-set: 17
Label: Code-Review=0