Update patch set 2
Patch Set 2:

I think this patch is good, but I wonder if it is consistent with the rest of the codebase?  

As far as I recall, all of the pixel processing in intra prediction/inter prediction/transform/loopfilter/dering filter produces output with the width/height rounded up to a multiple of 8x8 units.  (Perhaps the intention is to avoid edge cases in the hardware/simd implementation?)

On the other hand, the inter prediction does clip reference images to their exact clipped width/height.

Perhaps it would be worth seeking a wider consensus?

If you do still want to make this change, then it would make sense for the dering code to also change to use the cropped width.

Patch-set: 2
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=0