Update patch set 2
Patch Set 2:

Without this fix we get an encoder/decoder mismatch.  I agree that it will be good to harmonise how odd edges are handled throughout the code, but since "make test" currently fails, I'll submit this patch right now and check later if the code should be harmonised to agree with how it's done elsewhere.

Patch-set: 2
Label: Code-Review=0