Update patch set 2
Patch Set 2:

The extension to higher bit-depths for CLPF is straight-forward.  The operations are the same as for 8 bit, except that the strength has to be scaled to cap the same luminance changes.  The strengths for 8 bit is 1, 2 or 4, which for 10 bit become 4, 8 or 16, and for 12 bit they become 16, 32 or 64.

The RDO will shift the pixels down to 8 bit for deciding the strength and block size.  No quality loss was observed and this simplifies the SIMD code, as 16 pixels still can be processed in parallel using 128 bit wide registers, and code can be shared with the 8 bit implementation.  The actual filter uses the full resolution, of course.

Unit tests have been added to verify that the SIMD and C implementations produce identical results (--gtest_filter=*Clpf*).

Everything has been tested on x86 (sse2, ssse3 & sse4_1) and arm (neon).

Patch-set: 2
Label: Code-Review=0