Add bitstream debug tool
This is a debug tool used to detect bitstream error. On encoder side, it
pushes each bit and probability into a queue before the bit is written into
the Arithmetic coder. On decoder side, whenever a bit is read out from the
Arithmetic coder, it pops out the reference bit and probability from the
queue as well. If the two results do not match, this debug tool will report
an error.  This tool can be used to pin down the bitstream error precisely.
By combining gdb's backtrace method, we can detect which module causes the
bitstream error.

Change-Id: Iec6d4c2e3851b8ac1db35776fd7a216a9230a983