Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

Looks fine. Could use a comment on the condition change (<16), for sure need one in the change description.

Doesn't this apply to the RDO version as well?

Also, didn't notice this on the initial review, but the _ss*.c/_neon.c files should be in the x86 and arm directories, respectively. Not a huge deal, but would help keep things tidy.

Patch-set: 1
Label: Code-Review=+1