Update patch set 2
Patch Set 2:

> Well in that case it shouldn't be building any of the simd files,
> correct? Or this is when you want to force the C version of your
> simd code?

Correct.  If noe HAVE_SSE2 etc, there will be no simd code and the rtcd stuff will pick the C versions of the functions, so in order to get simd code with C implementations of the intrinsics, it's necessary to force it to do so (in aom_simd.h).  The C implementations of the intrinsics are intended for porting to new architectures and for debugging.  And if they're somehow accidentally activated, the unit tests will detect the slowdown.

Patch-set: 2