Update patch set 1
Patch Set 1:

I found a way to improve the 8 bit wide unsigned shifts.  Previously they were extended to 16 bit and then packed again, but it's better to mask out what's going to be shifted out, so we can use 16 bit (or whatever width) shift on the a 8 bit data.  This doesn't work for signed shift, though, but I cleaned up those a bit as well.

Patch-set: 1
Label: Code-Review=0