Update patch set 1
Patch Set 1:

Misc-fixes were taken by AOMedia as part of the baseline for AV1. This commit merely remove the #if CONFIG_MISC_FIXES #else #endif to cleanup the code around, so there should not be any impact at all. 

As to the original change of "always using 1/8 precision of motion vectors". I remember the impact on compression is around .1%, Debargha, does that sound close to you?

Patch-set: 1
Label: Code-Review=0