Update patch set 8
Patch Set 8:

Jenkins is hard to please.  The problem now seems to be that HAVE_SSE2 is defined, but the compiler runs without the -msse2 option or better, and then the compiler will complain about the use of intrinsics.  Do you know how the correct option can be added?  Apparently this is not a problem in the av1/ directory but is a problem in the test/ directory.  Or the problem could be that the proper -m option is added for C but not for C++.

Patch-set: 8