Update patch set 10
Patch Set 10:

> I think I know how to get it to work.  I need to get the Makefile
 > to contain rules like:
 > $(BUILD_PFX)%_sse2.cc.d: CXXFLAGS += -msse2 $(STACKREALIGN)
 > etc
 > just as it already does for .c files (I'm not sure how that is done
 > - by changing the configure script?)
 > 
 > Then I need to name my files in the test/ directory _sse2.cc etc,
 > one file for each architecture. This can be accomplished without
 > code duplication by using the preprocessor, similar to what is done
 > for CLPF.
 > 
 > I'll look into this tomorrow.  Meanwhile, if you know where I must
 > make changes to get the new rules in the Makefile, please let me
 > know.

They are manually added to the Makefile:
https://aomedia.googlesource.com/aom/+/master/build/make/Makefile#128

For example, this was for AVX[2]:
https://chromium.googlesource.com/webm/libvpx/+/e6863ef3185faf90e7be2c7f2c90bce0c08da1eb%5E%21/#F0

I'm still not sure this is the right approach. I don't think there is any particular reason to add C++ files that implement this. You could make stub C files, but you'd probably still end up with a pile of macros.

In either case, it's probably best to do a demonstration implementation with just a couple of functions so you don't burn too much time on something that won't work.

Patch-set: 10
Label: Code-Review=0