Update patch set 1
Patch Set 1:

Johann, can you +2 this trivial patch?  It fixes a problem that will break the build on 32 bit x86 with SSE4.1 if dotp intrinsics are used.  It turns out that _mm_cvtsi128_si64 isn't available on 32 bit systems, cf https://github.com/p12tic/libsimdpp/issues/1

I need to rebase #4560 on this patch in order to get #4560 to work (along with another upload of that one).

Patch-set: 1
Label: Code-Review=0