Update patch set 14
Patch Set 14:

> Success!
 > 
 > Johann, do the tests look ok?  It's not necessary to review the
 > actual 1000+ tests, just a quick look at the test framework and the
 > change to the build system.

 > Success!
 > 
 > Johann, do the tests look ok?  It's not necessary to review the
 > actual 1000+ tests, just a quick look at the test framework and the
 > change to the build system.

The design of the test still doesn't allow for building but not running a test.

This isn't much of an issue with up to sse4 because most processors support that. However, when you add avx/avx2 the tests will fail if the hardware does not support it. There is no runtime detecting in your test.

Patch-set: 14