Update patch set 32
Patch Set 32:

These tests turned out to be a hard nut.  The original strategy by which the test instantiated and called the intrinsic functions directly had a problem: The test had to be compiled with the proper -m setting, which could lead to illegal instruction even if the tests were filtered out because the hosting computer didn't support the architecture.  I suspect the reason is that some code gets run by the test class regardless of whether the test are run or not.

The solution was to separate the gtest code from the code actually referencing and running the intrinsic function.  The gtest code will only refer to the functions to be tested by name (a string) and this code will be compiled without extra -m flags.  Only if the tests are actually run, the code in the simd_test_arch_* files will be run and these files will be compiled with the proper -m setting.  These files will interpret the name (string) and call the proper functions.

Patch-set: 32
Label: Code-Review=0