Update patch set 33
Patch Set 33:

(2 comments)

> (6 comments)
 > 
 > simd_test_arch.h is a bit involved. can you bring some of this in
 > stages so it's a bit easier to review?

I don't think it makes sense to split the patch into stages, if that is what you meant, but I've added a few more comments.

The challenge with these tests is that there are several hundred functions to be tested and care was taken to do that in a compact way, but some wrappers seem unavoidable, which take up much space.

The key purpose of this file is to provide a way to compare these several hundred functions using intrinsics to corresponding reference functions written in plain C (the same name with a c_ prefix).  Templates, type equality and casting were used to avoid code duplication.  The result might be a bit hard to read, but the alternative duplicated code would be far worse.

Patch-set: 33
Label: Code-Review=0