Update patch set 38
Patch Set 38:

> Large changes like this tend to miss things
 > in review (even though that often require quite a few iterations)
 > increasing churn later.

This is already the first step of a couple of more commits.  The next is the addition of v256 intrinsics along with tests.  The tests for v64 and v128 could be split into two separate, but I chose not to because v64 and v128 are somewhat couple on x86 because they both use the m128i type, which had to be tackled specially.  So either a separate patch for v64 would include things that would not make sense for v64 alone, or the following v128 patch would undo much of what was done for v64.  Also, having a series of dependent patches makes it hard to rebase the dependent patches when code review cause changes to the first (I'm already struggling with that - have to other patches that I have to update every time I make changes to this patch).

Patch-set: 38
Label: Code-Review=0