Update patch set 10
Patch Set 10:

The 8-tap interpolation filter was used here in sub-pixel motion search instead of the bilinear filter originally used. This gave the gain. Up-sampling the reference frames was done to avoid to do filtering on the fly, which speeded up the search process. We realized the memory usage for large resolutions. For 4k encoding, We can disable this feature or do filtering on the fly if we want this quality gain. I am ok with either option.

Patch-set: 10