Update patch set 10
Patch Set 10:

OK, I can submit a patch to turn this off for resolutions >1080 high, which is an easy fix.

How much slower is doing the upsamping on the fly vs caching it in the upsampled reference frame?

There might be a better sweet spot for performance here, such as upsampling only 2x or 4x and using the approximation for the rest of the subpel levels. The current upsampling effectively decreases cache size by 64x.

Patch-set: 10