Update patch set 5
Patch Set 5:

> This patch preserves the current behavior. Ans had two experiments,
 > one with binary trees only and one with multisymbol (and still some
 > leftover binary(tree) symbols). The pre-existing daala_ec
 > experiment implicitly turned on multisymbol coding, so that's what
 > this change does.

Will there continue to be a distinction between the ANS and RANS experiments?  It seems you have replaced rans with ec_multisymbol and are now using ans to mean both the binary experiment and the multisymbol ranged coder.  Is it no longer possible to code binary trees with just ans?

Patch-set: 5