Update patch set 1
Patch Set 1:

Backwards updates of pdfs are still happening. At the end of the previous frame, the pdfs get updated by the backwards updates, and so have changed. The cdfs have adapted symbol by symbol and so have changed, but differently. This patch sets the cdfs from the pdfs before any encoding in the frame happens, so they are back in sync.

Before, the cdfs were being inherited from frame to frame independently of the backward PDF adaptation, and only synced in the forward update process. This happened in coefficient coding, where we sent some forward updates but not all, and then synced cdfs and pdfs. There was a corner case where is <20 symbols were coded, the encoder would not send an update and not sync, but the decoder would sync, hence a mismatch.

Patch-set: 1