Update patch set 1
Patch Set 1:

This appears to work: the output files are almost identical from rtcd.pl for cmake and make w/generic-gnu. The one difference is the include guards change for some reason. Didn't really pay attention to that because it's harmless (and the ones I generate are actually more correct).

ptal/ this still won't build, but i need more commits for that... trying to keep these smallish.

Patch-set: 1
Label: Code-Review=0