Update patch set 1
Patch Set 1:

This (to me) looks like it should work, and it does when invoked directly (the perl does, anyway). When cmake runs that's not the case, and I'm still looking into it. 

ptal at this, and I'll swing back and look at comments in the earlier patches first thing tomorrow. Thanks!

Also, please be critical of my first perl script... it really is the first one I've written.

Patch-set: 1
Label: Code-Review=0