Update patch set 1
Patch Set 1:

It looks like the nextgenv2 code is attempting to handle chroma for 4:2:2, but it's not doing it the correct way. In master, I've recently landed patches that handle 4:2:2 by simply not deringing chroma, which I think is more correct. I also made many changes to the deringing code that will need to be merged.

Patch-set: 1