Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> av1/encoder/encoder.c:53:30: fatal error: ./aom_scale_rtcd.h: No such file or directory
>  #include "./aom_scale_rtcd.h"

Can't repro this: Where'd you run it/how'd you make it break? No repro in or out of tree for me, but doesn't mean it's not broken...

One thing I did notice (and the other reason for the update to compiler flags setup): CMAKE_{C|CXX}_FLAGS was growing on every generation run. You might be in a bad state if you regenerated multiple times in the same build dir.

Patch-set: 2
Label: Code-Review=0