Update patch set 3
Patch Set 3:

> Change has been successfully rebased as 133c57c331f3b1806233fbd83db1f5f022b1490a
 > by Thomas Davies

This is from David in Argon:

Commit 133c57c3 is causing four test failures (AV1/AVxEncoderThreadTest.EncoderResultTest/{2,3,4,5}) on libaom with default settings, stating that the output differs between single-threaded and multi-threaded encodes.
 
As far as I can tell, what happens is that the bit writer buffer is not zeroed out before use. This was fine before the above patch (and on the current HEAD of nextgenv2), as each byte would effectively be zeroed out when the first bit of that byte is written to. However, after the patch, this does not happen. So if the total number of bits written is not a multiple of 8, some uninitialized bits are left in the last byte of the output stream.
 
This isn’t necessarily a problem, as it won’t affect the decoder at all. But the above tests expect to be able to compare two output streams byte-by-byte. I guess the decision which needs to be made is: Do we want the encoder to zero-pad the last byte written so that byte-by-byte comparisons are valid, or should any tests which want this behaviour be changed to explicitly zero out their buffers?

Patch-set: 3