Add a new experiment TRIPRED
Currently, the implementation of TRIPRED is trying to take advantage of
the existing wedge-based compound prediction. In stead of using a new
mask, the current scheme is to reuse all the mask options provided by
the existing wedge-based predictors. Following it a tri-inter-prediction
is proposed:

The bi-prediction by the existing compound mode is used to provide the
first predictor for the wedge-based inter prediction, and a third mv
using a third reference frame is estimated to provide the 2nd predictor
on the other side of the wedge.

A unique point is that the existing compound mode comes from an implicit
pair of reference frames, currently LAST/BWDREF, which would be further
extended. Hence the reference pair do not need to be written to the
bistream.

Change-Id: I49b4cd15f3af723c3bbe6015321a18a88b4d1648