Superframe index: Parse correct number of frame sizes
If there are N frames in a superframe, the superframe
index will enumerate the sizes of only the first
(N - 1) frames. The code assumed that all N frame
sizes are coded.

The size of the final frame in the superframe is
calculated as follows:
(data_size - index_size - sum_of_N-1_frame_sizes)

Change-Id: Ia53ae27455a4f024eec311bf7356a00e8e0d449a