Update patch set 3
Patch Set 3:

> Wondering if a better fix would be to make sure consistently in the
 > encoder and the decoder that the motion vector assigned to ZEROMV
 > with global motion used is obtained from gm_get_motion_vector(...).
I checked that when submitting my previous patch, and it looked like everything was consistent. This patch is to address a different case, where the global motion vector is nonzero but all candidate MVs are zero. In that case, setting the ALL_ZERO flag means we can't use NEARMV/NEARESTMV to encode a zero motion vector, even if we would otherwise be able to.

Unfortunately, the nice solution of comparing to the global motion vector instead of zero doesn't work: in the encoder, we call av1_find_mv_refs only in the rdopt phase, well before the last check to see if we want to throw away the global motion model (in write_global_motion_params). So that would lead to inconsistencies.

Patch-set: 3
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=0